Stream buffering, faster writes & state transitions
Stream buffering improves protocol compliance and should correctly
handle pipelined incoming requests (although backend requests are not
pipelined). Writes no longer have to wait for an epoll call; they'll be
attempted immediately and fall back to an epoll-triggered write if the
socket wasn't ready or there was too much data to send in one go.

State transitions from read-to-write can now happen immediately rather
than going via an epoll to wait for a ready file descriptor.

Jobs are no longer freed during close_job() in case there are multiple
file descriptors being waited for in the same job. This is not
something that currently happens (we only subscribe to either the
backend or frontend/client fd), although it'll probably be necessary to
implement timeouts, because multiple event sources could be live &
trigger for the same job.

There are still some bugs that prevent this from working (at all).